+++ PortAIO.csproj
@@ -1499,6 +1499,19 @@
     <Compile Include="Core\Champion Ports\LeBlanc\LeBlanc II\Modes\ModePerma.cs" />
     <Compile Include="Core\Champion Ports\LeBlanc\LeBlanc II\Modes\ModeSettings.cs" />
     <Compile Include="Core\Champion Ports\LeBlanc\LeBlanc II\Program.cs" />
+    <Compile Include="Core\Champion Ports\Lee Sin\ElLeeSin\Components\MyMenu.cs" />
+    <Compile Include="Core\Champion Ports\Lee Sin\ElLeeSin\Components\SpellManagers\EscapeHandler.cs" />
+    <Compile Include="Core\Champion Ports\Lee Sin\ElLeeSin\Components\SpellManagers\InsecManager.cs" />
+    <Compile Include="Core\Champion Ports\Lee Sin\ElLeeSin\Components\SpellManagers\PassiveManager.cs" />
+    <Compile Include="Core\Champion Ports\Lee Sin\ElLeeSin\Components\SpellManagers\Wardmanager.cs" />
+    <Compile Include="Core\Champion Ports\Lee Sin\ElLeeSin\Components\Spells\SpellE.cs" />
+    <Compile Include="Core\Champion Ports\Lee Sin\ElLeeSin\Drawings.cs" />
+    <Compile Include="Core\Champion Ports\Lee Sin\ElLeeSin\Kurisu\BubbaSharp.cs" />
+    <Compile Include="Core\Champion Ports\Lee Sin\ElLeeSin\LeeSin.cs" />
+    <Compile Include="Core\Champion Ports\Lee Sin\ElLeeSin\Program.cs" />
+    <Compile Include="Core\Champion Ports\Lee Sin\ElLeeSin\Utilities\Geometry.cs" />
+    <Compile Include="Core\Champion Ports\Lee Sin\ElLeeSin\Utilities\ItemManager.cs" />
+    <Compile Include="Core\Champion Ports\Lee Sin\ElLeeSin\Utilities\Misc.cs" />
     <Compile Include="Core\Champion Ports\Lee Sin\LeeSin EloClimber\Combo.cs" />
     <Compile Include="Core\Champion Ports\Lee Sin\LeeSin EloClimber\Insec.cs" />
     <Compile Include="Core\Champion Ports\Lee Sin\LeeSin EloClimber\Jungle.cs" />
@@ -1789,14 +1802,6 @@
     <Compile Include="Core\Champion Ports\Warwick\Warwick ll\Modes\ModeSpellW.cs" />
     <Compile Include="Core\Champion Ports\Warwick\Warwick ll\Program.cs" />
     <Compile Include="Core\Champion Ports\Warwick\Warwick ll\Warwick.cs" />
-    <Compile Include="Core\Champion Ports\Lee Sin\ElLeeSin\Drawings.cs" />
-    <Compile Include="Core\Champion Ports\Lee Sin\ElLeeSin\EscapeHandler.cs" />
-    <Compile Include="Core\Champion Ports\Lee Sin\ElLeeSin\InitMenu.cs" />
-    <Compile Include="Core\Champion Ports\Lee Sin\ElLeeSin\Program.cs" />
-    <Compile Include="Core\Champion Ports\Lee Sin\ElLeeSin\Utilities\Geometry.cs" />
-    <Compile Include="Core\Champion Ports\Lee Sin\ElLeeSin\Utilities\Misc.cs" />
-    <Compile Include="Core\Champion Ports\Lee Sin\ElLeeSin\Utilities\PassiveManager.cs" />
-    <Compile Include="Core\Champion Ports\Lee Sin\ElLeeSin\Utilities\Wardmanager.cs" />
     <Compile Include="Core\Champion Ports\Varus\ElVarus\DrawDamage.cs" />
     <Compile Include="Core\Champion Ports\Varus\ElVarus\Drawings.cs" />
     <Compile Include="Core\Champion Ports\Varus\ElVarus\ElVarusMenu.cs" />
